update contracts + setAdmin (started)

+++ src/indexer/mavryk/handlers/on_doorman_update_min_mvk_amount.py
@@ -11,10 +11,10 @@ async def on_doorman_update_min_mvk_amount(
 ) -> None:
     # Get doorman contract
     doorman_address = update_min_mvk_amount.data.target_address
-    doorman = await models.Doorman.get(address=doorman_address)
+    doorman         = await models.Doorman.get(address=doorman_address)
 
     # Get new minimum value
-    new_min_amount = int(update_min_mvk_amount.parameter.__root__)
+    new_min_amount  = float(update_min_mvk_amount.parameter.__root__)
     
     # Save the new value
     doorman.min_mvk_amount = new_min_amount
